rebase-fix: done to fix the rebase issues

+++ xlators/storage/posix/src/posix.h
@@ -26,22 +26,22 @@
 } while (0);
 
 struct posix_private {
-  int32_t temp;
+  int temp;
   char is_stateless;
   char is_debug;
   char base_path[PATH_MAX];
-  int32_t base_path_length;
+  int base_path_length;
 
   struct xlator_stats stats; /* Statastics, provides activity of the server */
   
   struct timeval prev_fetch_time;
   struct timeval init_time;
-  int32_t max_read;            /* */
-  int32_t max_write;           /* */
-  int64_t interval_read;      /* Used to calculate the max_read value */
-  int64_t interval_write;     /* Used to calculate the max_write value */
-  int64_t read_value;    /* Total read, from init */
-  int64_t write_value;   /* Total write, from init */
+  int max_read;            /* */
+  int max_write;           /* */
+  long interval_read;      /* Used to calculate the max_read value */
+  long interval_write;     /* Used to calculate the max_write value */
+  long long read_value;    /* Total read, from init */
+  long long write_value;   /* Total write, from init */
 };
 
 #endif /* _POSIX_H */
